# openapi.yaml
openapi: 3.0.3
info:
  title: HackNext REST API
  description: Offline-first production hackathon platform API for collegiate engineering sprints.
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
    description: Local self-hosted backend server
paths:
...
  /judge/scores:
    get:
      summary: Get assigned project scores for authenticated judge
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Assigned submissions and existing ratings
        '403':
          description: Forbidden if attempting to query peer judge scores
  /judge/scores/peer:
    get:
      summary: Peer score access endpoint (Strictly forbidden for judges)
      security:
        - BearerAuth: []
      responses:
        '403':
          description: Forbidden (Strict role isolation)
...
  /export/backup/full:
    get:
      summary: Export full database state as an atomic JSON snapshot
      security:
        - BearerAuth: []
      responses:
        '200':
          description: JSON database snapshot
  /export/backup/restore:
    post:
      summary: Restore entire database atomically from JSON backup
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Database restored successfully
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT